staging/pkg/mod/golang.org/x/net@v0.0.0-20201110031124-69a78807bb2b/html/testdata/webkit/tests17.dat

#data
<!doctype html><table><tbody><select><tr>
#errors
(1,37): unexpected-start-tag-implies-table-voodoo
(1,41): unexpected-table-element-start-tag-in-select-in-table
(1,41): eof-in-table
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|     <table>
|       <tbody>
|         <tr>

#data
<!doctype html><table><tr><select><td>
#errors
(1,34): unexpected-start-tag-implies-table-voodoo
(1,38): unexpected-table-element-start-tag-in-select-in-table
(1,38): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|     <table>
|       <tbody>
|         <tr>
|           <td>

#data
<!doctype html><table><tr><td><select><td>
#errors
(1,42): unexpected-table-element-start-tag-in-select-in-table
(1,42): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>
|           <td>

#data
<!doctype html><table><tr><th><select><td>
#errors
(1,42): unexpected-table-element-start-tag-in-select-in-table
(1,42): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <th>
|             <select>
|           <td>

#data
<!doctype html><table><caption><select><tr>
#errors
(1,43): unexpected-table-element-start-tag-in-select-in-table
(1,43): eof-in-table
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <select>
|       <tbody>
|         <tr>

#data
<!doctype html><select><tr>
#errors
(1,27): unexpected-start-tag-in-select
(1,27): eof-in-select
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>

#data
<!doctype html><select><td>
#errors
(1,27): unexpected-start-tag-in-select
(1,27): eof-in-select
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>

#data
<!doctype html><select><th>
#errors
(1,27): unexpected-start-tag-in-select
(1,27): eof-in-select
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>

#data
<!doctype html><select><tbody>
#errors
(1,30): unexpected-start-tag-in-select
(1,30): eof-in-select
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>

#data
<!doctype html><select><thead>
#errors
(1,30): unexpected-start-tag-in-select
(1,30): eof-in-select
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>

#data
<!doctype html><select><tfoot>
#errors
(1,30): unexpected-start-tag-in-select
(1,30): eof-in-select
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>

#data
<!doctype html><select><caption>
#errors
(1,32): unexpected-start-tag-in-select
(1,32): eof-in-select
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>

#data
<!doctype html><table><tr></table>a
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|     "a"
